platform: msm: gsi: Correct debug argument length checks.
These preliminary argument length checks were added to prevent
allocating memory if the argument is going to be invalid anyway, but
didn't account for cases that take a single argument of length 1, or two
arguments separated with a space of at least length 3.
Update the checks to handle these cases.

Change-Id: Ie18d8239020fe281f4cb6b63eb29ff69dd60964c
Signed-off-by: MarijnS95 <marijns95@gmail.com>