init: initramfs: Add config option to change the skip_initramfs bp name
Some bootloaders will set this boot parameter whenever a direct
boot from a newly uploaded-to-RAM boot image is requested.
Sometimes though we want it to load the initramfs anyway, even
if it's a direct boot.
This configuration option will rename the "skip_initramfs" to
"do_skip_initramfs".

This is a kernel hack. Definitely say N.

init: Kconfig: Add decription to DEBUG_RENAME_SKIP_INITRAMFS_BOOTPARAM

Add a description to the configuration option.